+++ database/migrations/2024_06_10_033800_create_servicios_table.php
@@ -8,24 +8,30 @@ return new class extends Migration
 {
     /**
      * Run the migrations.
+     * 
+     * @return void
      */
     public function up(): void
     {
-        Schema::create('servicios', function (Blueprint $table) {
-            $table->id();
-            $table->string('nombre');
-            $table->text('descripcion');
-            $table->decimal('precio', 8, 2);
-            $table->unsignedBigInteger('medico_id')->nullable();
-            $table->timestamps();
-
-            // Crear llave foranea que extraiga de la tabla de users aquellos que tengan el rol de médico
-            $table->foreign('medico_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if(!Schema::hasTable('servicios')) {
+            Schema::create('servicios', function (Blueprint $table) {
+                $table->id();
+                $table->string('nombre');
+                $table->text('descripcion');
+                $table->decimal('precio', 8, 2);
+                $table->unsignedBigInteger('medico_id')->nullable();
+                $table->timestamps();
+    
+                // Crear llave foranea que extraiga de la tabla de users aquellos que tengan el rol de médico
+                $table->foreign('medico_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
+     * 
+     * @return void
      */
     public function down(): void
     {
